feat(encoder): encoder 컴포넌트화, 적용
